% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilasso.R
\name{.ilasso2}
\alias{.ilasso2}
\title{Internal ilasso function}
\usage{
.ilasso2(x, y, hierarchy, maineffects1, family, standardize, reruns, ...)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format.}

\item{y}{response variable. Quantitative for family="gaussian". For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; for a factor, the last level in alphabetical order is the target class). For "binomial" if y is presented as a vector, it will be coerced into a factor.}

\item{hierarchy}{character. The hierarchy of the model.}

\item{maineffects1}{numeric vector. This is for the \code{step = "second"}, and allows the user to provide the main effects for which interactions are included (or not, depending on the \code{hierarchy}).}

\item{family}{Response type (see above).}

\item{standardize}{logical. Should x be standardised prior to fitting the model. Defaults to \code{TRUE}.}

\item{reruns}{numeric. The number of reruns of the procedure, the default is 10.}

\item{...}{Other arguments passed to the cv.glmnet function.}
}
\description{
Internal ilasso function
}
