% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milasso.R
\name{.milasso2}
\alias{.milasso2}
\title{Internal milasso function}
\usage{
.milasso2(
  X,
  Y,
  hierarchy,
  preserve,
  X1,
  Y1,
  standardize,
  fold,
  seed,
  reruns,
  ...
)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format.}

\item{Y}{response matrix, of dimension nobs x nouts. Quantitative for family="gaussian".}

\item{hierarchy}{character. The hierarchy of the model.}

\item{X1}{numeric vector. This is for the \code{step = "second"}, and allows the user to provide the main effects for which interactions are included (or not, depending on the \code{hierarchy}).}

\item{Y1}{numeric vector. This is for the \code{step = "second"}, and allows the user to provide the responses for which interactions are included (or not, depending on the \code{hierarchy}).}

\item{standardize}{logical. Should X be standardised prior to fitting the model. Defaults to \code{TRUE}.}

\item{fold}{a positive integer for the corss validation fold. Default is \code{fold = 5}.}

\item{seed}{a numeric scaler, specifying the seed of the random number generator in R for generating cross validation subset for each fold. Default=1.}

\item{reruns}{numeric. The number of reruns of the procedure, the default is 10.}

\item{...}{Other arguments passed to the MSGLasso.cv function.}
}
\description{
Internal milasso function
}
