% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilasso.R
\name{ilasso}
\alias{ilasso}
\title{Identify interactions with lasso}
\usage{
ilasso(
  x,
  y,
  step = c("both", "first", "second"),
  hierarchy = c("strong", "weak"),
  maineffects1,
  family = c("gaussian", "binomial"),
  standardize = TRUE,
  reruns = 10,
  ...
)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format.}

\item{y}{response variable. Quantitative for family="gaussian". For family="binomial" should be either a factor with two levels, or a two-column matrix of counts or proportions (the second column is treated as the target class; for a factor, the last level in alphabetical order is the target class). For "binomial" if y is presented as a vector, it will be coerced into a factor.}

\item{step}{character. Which step should be run. For \code{step = "both"} or \code{step = "second"}, a value for \code{hierarchy} must also be provided. For \code{step = "second"}, values for \code{maineffects1} must also be provided.}

\item{hierarchy}{character. The hierarchy of the model.}

\item{maineffects1}{numeric vector. This is for the \code{step = "second"}, and allows the user to provide the main effects for which interactions are included (or not, depending on the \code{hierarchy}).}

\item{family}{Response type (see above).}

\item{standardize}{logical. Should x be standardised prior to fitting the model. Defaults to \code{TRUE}.}

\item{reruns}{numeric. The number of reruns of the procedure, the default is 10.}

\item{...}{Other arguments passed to the cv.glmnet function.}
}
\value{
An object with S3 class "ilasso". Selected main effects and interactions. \item{maineffects1 }{For \code{step = "both"} or \code{step = "first"} only. A vector consisting of the indecies of the main effects selected in step 1.} \item{maineffects1 }{For \code{step = "both"} or \code{step = "second"} only. A vector consisting of the indecies of the main effects selected in step 2.} \item{interactions2 }{For \code{step = "both"} or \code{step = "second"} only. A matrix consisting of the indecies of the main effects (columns) corresponding to the pairwise interactions (rows) selected in step 2.}
}
\description{
Identifies main effects and pairwise interactions in a linear or logistic regression model via hierarchical lasso penalisation. Can deal with all shapes of data, including very large sparse data matrices.
}
\details{
Step 1 assumes the the pure main effects model and returns the main effects selected by lasso corresponding to the set of non-zero estimates.

Step 2 assumes a pairwise interaction model subject to a hierarchical restriction and returns the main effects and interactions selected by the adaptive lasso corresponding to the set of non-zero estimates.
}
\examples{
# Gaussian
p <- 200
n <- 100
nz <- dplyr::arrange(expand.grid(a = 1:p, b = 1:p), a)
X <- matrix(sample(0:2, n*p, replace=TRUE), nrow=n, ncol=p)
ind <- t(combn(ncol(X),2))
out <- apply(ind, 1, function(x) X[,x[1]] * X[,x[2]])
XX <- cbind(X, out)
beta <- c(3,3, rep(0, ncol(X)-2))
beta12 <- c(3, rep(0, choose(p, 2)-1))
y <- XX \%*\% c(beta,beta12) + rnorm(n)
# Step 1
fit1 <- ilasso(x = X, y = y, step = "first")
me1 <- fit1$maineffects1
print(fit1)
# Step 2
fit2 <- ilasso(x = X, y = y, step = "second", maineffects1 = me1)
print(fit2)

}
\seealso{
\code{\link{glmnet}} and \code{\link{cv.glmnet}}
}
