% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ilasso.R
\name{predict.ilasso}
\alias{predict.ilasso}
\title{Extract coefficients from an ilasso object}
\usage{
\method{predict}{ilasso}(object, newx, ...)
}
\arguments{
\item{object}{Fitted "ilasso" object.}

\item{newx}{Matrix of new values for x of dimension nobs x nouts.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
The object returned is a vector of length nobs of the fitted values.
}
\description{
Similar to other predict methods, this function predicts fitted values and extract coefficients from a fitted "edgwas" object.
}
\examples{
# Gaussian
p <- 20
n <- 100
nz <- dplyr::arrange(expand.grid(a = 1:p, b = 1:p), a)
X <- matrix(sample(0:2, n*p, replace=TRUE), nrow=n, ncol=p)
ind <- t(combn(ncol(X),2))
out <- apply(ind, 1, function(x) X[,x[1]] * X[,x[2]])
XX <- cbind(X, out)
beta <- c(3,3, rep(0, ncol(X)-2))
beta12 <- c(3, rep(0, choose(p, 2)-1))
y <- XX \%*\% c(beta,beta12) + rnorm(n)
# Fit weak hierarchy
fit <- ilasso(x = X, y = y, hierarchy = "weak")
# Predict
newx <- matrix(sample(0:2, n*p, replace=TRUE), nrow=n, ncol=p)
yHat <- predict(fit, newx = newx)

}
